+++ src/java/org/apache/cassandra/db/compaction/TimeWindowCompactionTask.java
@@ -20,7 +20,6 @@ package org.apache.cassandra.db.compaction;
 
 import java.util.Set;
 
-import org.apache.cassandra.db.ColumnFamilyStore;
 import org.apache.cassandra.db.lifecycle.LifecycleTransaction;
 import org.apache.cassandra.io.sstable.format.SSTableReader;
 
@@ -28,7 +27,7 @@ public class TimeWindowCompactionTask extends CompactionTask
 {
     private final boolean ignoreOverlaps;
 
-    public TimeWindowCompactionTask(TimeWindowCompactionStrategy strategy, LifecycleTransaction txn, int gcBefore, boolean ignoreOverlaps)
+    public TimeWindowCompactionTask(TimeWindowCompactionStrategy strategy, LifecycleTransaction txn, long gcBefore, boolean ignoreOverlaps)
     {
         super(strategy, txn, gcBefore, false);
         this.ignoreOverlaps = ignoreOverlaps;
